Fix multiple displays by running a Form on each one
The previous implementation, taken from a C# screensaver tutorial,
used Application.Run to launch one screensaver form on each display.
However this meant they were run in sequence instead of at the same
time. Now we wrap the forms in an ApplicationContext to run them in
parallel.

The new code is based on this SO answer: https://stackoverflow.com/a/13406508